feat: migrate to Gatsby
Took 43 minutes
