+++ .prettierignore
@@ -1,0 +1,4 @@
+.cache
+package.json
+package-lock.json
+public
