--- components/ui/Logo.tsx
@@ -1,3 +1,0 @@
-import styled from "styled-components";
-
-export const Logo = styled.img``;
