--- components/ui/Title.tsx
@@ -1,20 +1,0 @@
-import styled from "styled-components";
-
-export const Title = styled.h2`
-  font-family: Montserrat;
-  font-size: 38px;
-  font-weight: 700;
-  line-height: 48px;
-  text-transform: uppercase;
-  letter-spacing: 6.08px;
-  color: #fff;
-  margin: auto;
-  &:before {
-    content: "";
-    display: block;
-    width: 30px;
-    height: 4px;
-    margin-bottom: 30px;
-    background-color: #fff;
-  }
-`;
