--- components/ui/globalStyle.tsx
@@ -1,9 +1,0 @@
-import { createGlobalStyle } from "styled-components";
-import theme from "../theme";
-
-export const GlobalStyle = createGlobalStyle`
-  body {
-    margin: ${theme.atoms.layout.desktop.margin};
-    font-family: Montserrat;
-  }
-`;
