+++ gatsby-config.js
@@ -1,0 +1,9 @@
+/**
+ * Configure your Gatsby site with this file.
+ *
+ * See: https://www.gatsbyjs.org/docs/gatsby-config/
+ */
+
+module.exports = {
+  plugins: ["gatsby-plugin-typescript"],
+}
