+++ package.json
@@ -1,35 +1,38 @@
 {
-  "name": "site-2020",
-  "version": "1.0.0",
-  "description": "New site home page",
-  "main": "index.js",
+  "name": "gatsby-starter-hello-world",
+  "private": true,
+  "description": "A simplified bare-bones starter for Gatsby",
+  "version": "0.1.0",
+  "license": "MIT",
   "scripts": {
-    "start:dev": "next",
-    "build": "next build",
-    "start": "next start"
+    "build": "gatsby build",
+    "develop": "gatsby develop",
+    "format": "prettier --write \"**/*.{js,jsx,json,md}\"",
+    "start": "npm run develop",
+    "serve": "gatsby serve",
+    "clean": "gatsby clean",
+    "test": "echo \"Write tests! -> https://gatsby.dev/unit-testing\" && exit 1"
+  },
+  "dependencies": {
+    "@fullpage/react-fullpage": "^0.1.17",
+    "@material-ui/core": "^4.9.3",
+    "gatsby": "^2.19.7",
+    "gatsby-plugin-typescript": "^2.1.27",
+    "react": "^16.12.0",
+    "react-dom": "^16.12.0",
+    "styled-components": "^5.0.1"
+  },
+  "devDependencies": {
+    "@types/react": "^16.9.21",
+    "@types/react-dom": "^16.9.5",
+    "@types/styled-components": "^5.0.0",
+    "prettier": "^1.19.1"
   },
   "repository": {
     "type": "git",
-    "url": "git+ssh://git@github.com/desirelabs/site-2020.git"
+    "url": "https://github.com/gatsbyjs/gatsby-starter-hello-world"
   },
-  "author": "Franck LEBAS & Laura AUPITRE",
-  "license": "UNLICENCED",
   "bugs": {
-    "url": "https://github.com/desirelabs/site-2020/issues"
-  },
-  "homepage": "https://github.com/desirelabs/site-2020#readme",
-  "dependencies": {
-    "@fullpage/react-fullpage": "^0.1.17",
-    "@material-ui/core": "^4.9.0",
-    "next": "^9.2.1",
-    "react": "^16.12.0",
-    "react-dom": "^16.12.0",
-    "styled-components": "^5.0.0"
-  },
-  "devDependencies": {
-    "@types/node": "^13.7.0",
-    "@types/styled-components": "^4.4.2",
-    "babel-plugin-styled-components": "^1.10.7",
-    "typescript": "^3.7.5"
+    "url": "https://github.com/gatsbyjs/gatsby/issues"
   }
 }
