+++ src/components/ui/Logo.tsx
@@ -1,0 +1,3 @@
+import styled from "styled-components"
+
+export const Logo = styled.img``
