+++ src/components/ui/globalStyle.tsx
@@ -1,0 +1,9 @@
+import { createGlobalStyle } from "styled-components"
+import theme from "../theme"
+
+export const GlobalStyle = createGlobalStyle`
+  body {
+    margin: ${theme.atoms.layout.desktop.margin};
+    font-family: Montserrat;
+  }
+`
